pyproject.toml
[project]
name = "agentic-ai-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.22",
    "langchain>=1.2.7",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "starlette>=0.52.1",
    "uvicorn>=0.40.0",
]
